requirements.txt
greenlet==2.0.2
psycopg2-binary==2.9.6
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
SQLAlchemy==2.0.13
TgCrypto==1.2.5
typing_extensions==4.5.0
uvloop==0.17.0
